Work in progress on making the code work on non-GNU systems more sanely
This works (with warnings) on my system, but it probably doesn't work on a
non-GNU system and I'm not sure what's going on - removing the _GNU_SOURCE
declaration from maxminddb.h causes all sorts of breakage.

I need to better understand how these _X_SOURCE macros are supposed to work.